Allow passing custom exception message for authorize!
Sometimes, there is a need for specifying a message along with
authorization of a action. This message is useful as it provides a great
feedback to the consumer of the action. Plus, it acts as a documentation
which makes it clearer as to why this authorization was introduced in
some cases.

So, the message passed along with `authorize!` will be accessible on the
exception object if the authorization fails.